Update content_filters.py
Now we get parameters for each unique chat by id

+++ handlers/content_filters.py
@@ -5,6 +5,7 @@ from keyboards.toggles import anim_inline
 from aiogram import F
 from utils.neuro.stt import STT
 from decouple import config
+from models import database
 import os
 from pathlib import Path
 import main
@@ -23,26 +24,28 @@ async def toggle_animated_stickers(message: Message):
 
 @router.message(F.sticker)
 async def check_sticker(message: Message):
-    if not main.animations_allowed:
+    chat = database.Chat(message.chat.id).get()
+    if int(chat["anim"]):
         if message.sticker and (message.sticker.is_video or message.sticker.is_animated):
             await message.delete()
 
 
 @router.message(F.voice)
 async def voice_to_text(message: Message):
-    file_id = message.voice.file_id
+    chat = database.Chat(message.chat.id).get()
+    if int(chat["voice"]):
+        file_id = message.voice.file_id
 
-    file = await main.bot.get_file(file_id)
-    file_path = file.file_path
-    file_on_disk = Path(config("MEDIA_FULL_PATH") + "audio/", f"{file_id}.ogg")
-    await main.bot.download_file(file_path, destination=file_on_disk)
-    await message.reply("Аудио получено")
+        file = await main.bot.get_file(file_id)
+        file_path = file.file_path
+        file_on_disk = Path(config("MEDIA_FULL_PATH") + "audio/", f"{file_id}.ogg")
+        await main.bot.download_file(file_path, destination=file_on_disk)
+        await message.reply("Аудио получено")
 
-    text = stt.audio_to_text(file_on_disk)
-    print(file_on_disk)
-    if not text:
-        text = "Формат документа не поддерживается"
-    os.remove(file_on_disk)
-    await message.reply(text + ".")
-
+        text = stt.audio_to_text(file_on_disk)
+        print(file_on_disk)
+        if not text:
+            text = "Формат документа не поддерживается"
+        os.remove(file_on_disk)
+        await message.reply(text + ".")
